WIP refactoring code units
