+++ assets/terminal12x12_gs_ro.png
