+++ vfl/program.py
@@ -1,0 +1,4 @@
+class Program:
+
+    def __init__(self, parse_node):
+        self._node = parse_node
